+++ NEWS.md
@@ -1,3 +1,5 @@
+# withr (development version)
+
 # withr 2.4.1
 
 - Tests which require `capabilities("cairo")` are now skipped.