+++ NEWS.md
@@ -1,5 +1,7 @@
 # withr (development version)
 
+- `with_timezone()` now works after recent changes to `Sys.timezone()` in R-devel (#165)
+
 - `with_environment()` now passing `pos` and `warn.conflicts` to `attach()`, as intended (#161)
 
 # withr 2.4.1