+++ NEWS.md
@@ -5,7 +5,9 @@
 - `with_seed()` now also sets the RNG via new arguments `.rng_kind`, `.rng_normal_kind` and `.rng_sample_kind`
   (#162, @AshesITR).
 
-# withr 2.4.1
+- `with_timezone()` now works after recent changes to `Sys.timezone()` in R-devel (#165)
+
+withr 2.4.1
 
 - Tests which require `capabilities("cairo")` are now skipped.